fixed some signals between host and player

+++ host.c
@@ -1,3 +1,4 @@
+
 #include "host.h"
 
 #define MAX_PLAYERS 3
@@ -23,8 +24,6 @@ to the next question.
 3. the host loops through this array of pipes to know which pipe to send a question to
 
 CURRENT PROBLEMS: 
-- doesnt loop back to do quesitons after 3
-- only takes in the first 3 characteres of the answer
 - exiting the host doesnt make the player exit
 - lowercase for answers (not case sensitive, also add in instructions)
 */
@@ -49,25 +48,6 @@ struct player_struct create_player(char* player_num){
   return p;
 }
 
-// for pipe, look for sigpipe
-// SIGNAL HANDLING
-static void sighandler(int signo){
-    if (signo == SIGINT){
-      printf("\nGame disconnected. Players will disconnect automatically.\n");
-      for(int i = 0; i<MAX_PLAYERS; i++){
-        if(players[i].pid>0) kill(players[i].pid, SIGINT); // kills using pid
-      }
-      delete_pipes();
-      remove(WKP);
-      exit(0);
-    }
-    if (signo == SIGPIPE){
-      printf("\nDisconnected pipe.\n");
-		  delete_pipes();
-		  exit(0);
-    }
-}
-
 void print_points(){
   printf("FINAL POINT COUNT:\n");
     for (int i = 0; i<MAX_PLAYERS; i++){
@@ -80,14 +60,35 @@ void delete_pipes(){
   for (int i = 0; i<num_players; i++){
 			int pp = open(players[i].pipe_name, O_WRONLY);
 	  //sends a message so the players disconnect
+			printf("im deleting pipes\n");
 			char end[100] = "end";
 			write(pp, end, sizeof(end));
-      kill(players[i].pid,SIGINT); // send sigint
       unlink(players[i].pipe_name);
 	    remove(players[i].pipe_name);
+			unlink(WKP);
+			remove(WKP);
   }
-  unlink(WKP);
-	remove(WKP);
+}
+
+// for pipe, look for sigpipe
+// SIGNAL HANDLING
+static void sighandler(int signo){
+  if (signo == SIGINT){
+    printf("\nDisconnected game. Players will be disconnected automatically.\n");
+    for (int i = 0; i<MAX_PLAYERS; i++){
+      if (players[i].pid>0){
+        kill(players[i].pid,SIGINT);
+      }
+    }
+    delete_pipes();
+    remove(WKP);
+    exit(0);
+  }
+  if (signo == SIGPIPE){
+    printf("\nDisconnected pipe.\n");
+		delete_pipes();
+		exit(0);
+  }
 }
 
 // handles flow of the game, forking(?)
@@ -150,6 +151,14 @@ int main(){
       find_question(topic, question, answer);
 				//printf("this is the answer from the while loop: %s\n", answer);
 
+      // if it ran out of questions, say that and then break the loop to end the game
+      if(strlen(question)==1){
+        printf("No more questions! Game over."); // separate display points function
+        print_points();
+					delete_pipes();
+        break;
+      }
+
       // send question and answer to player through pipe!
       int send_q = open(players[curr_player].pipe_name,O_WRONLY);
       if (send_q < 0){
@@ -167,7 +176,7 @@ int main(){
 				printf("answer: %s\n", answer);
       close(send_q);
 
-      // now wait for answer...
+      // no wait for answer...
       int get_a = open(players[curr_player].pipe_name,O_RDONLY);
         if (get_a < 0){
 						delete_pipes();
@@ -179,10 +188,10 @@ int main(){
       // ok so we have to make it so that the player pipe writing side will send in a stdin input
       if(read(get_a,player_answer,sizeof(player_answer))>0){
         player_answer[strcspn(player_answer, "\n")] = '\0';
-
+				printf("received!: %s\n", player_answer);
+        // remove trailing newline here i forgot how
         if (strcmp(player_answer,"end")==0){
           printf("Player ended the game.\n");
-          delete_pipes();
           break; // point function
         }
         else if (strcmp(player_answer,answer)==0){
@@ -225,13 +234,6 @@ void find_question(char * topic, char* question, char* answer) {
 	}
 
 	char line[1000];
-  // if it ran out of questions, say that and end the game
-  if(fgets(line,sizeof(line),readfile)==NULL){
-    printf("No more questions! Game over."); // separate display points function
-    print_points();
-		delete_pipes();
-    // exit how..????
-  }
 	if (strcmp(topic, "history") == 0) {
 		for (int i = 0; i <= histq_num; i++) {
 			fgets(line, sizeof(line), readfile);
